pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "obs-remote"
version = "1.0.0"
description = "Advanced OBS Smart Controller"
readme = "README.md"
requires-python = ">=3.7"
authors = [
    {name = "Your Name", email = "your.email@example.com"},
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
dependencies = [
    "evdev",
    "obs-websocket-py",
    "psutil",
]

[project.scripts]
obs-remote = "main:main_cli"

[project.urls]
Homepage = "https://github.com/yourusername/obs-remote"

[tool.hatch.build.targets.wheel]
packages = ["main.py"]
